Fix non-Youtube player test
The default player doesn’t handle a non Youtube URL unless it has a
playable HTML5 media type.

The old test passed in `media: undefined` (the ID lookup from a Vimeo
URL never worked). which couldn’t be found in the compatible list of
types. When the supplied URL correctly gets past the "is this Youtube?"
check, the player tries to create an HTML5 media player without a valid
file/mime type and errors.

There's no concept of a Vimeo player in the TestRunner or specs, those
sources files aren't included.

* Refactor test to be a generic "non-Youtube" test with a filetype